requirements.txt
# Flask và extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.7

# Machine Learning
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# OpenAI
openai==1.3.0

# Environment
python-dotenv==1.0.0

# Authentication (giả sử bạn có module này)
# flask-auth==1.0.0